% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_and_remove_segments.R
\name{add_segments}
\alias{add_segments}
\title{Function for adding a segment from the chromosome fasta sequence}
\usage{
add_segments(chromosome_string, gain_segments)
}
\arguments{
\item{chromosome_string}{Chromosome fasta sequence string (one contiguous string)}

\item{gain_segments}{A two column dataframe with headers startpos and endpos providing regions to be simulated for GAIN}
}
\description{
Non-piecewise (legacy) function to add segment due to GAIN at the end of a chromosome fasta sequence based on start and end positions (using alternate_segments)
}
\author{
naser.ansari-pour
}
\keyword{internal}
