% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_and_merge_bams.R
\name{clone_proportion}
\alias{clone_proportion}
\title{Deprecated - Function for calculating clone proportion for normal contamination in non-pure simulated tumour}
\usage{
clone_proportion(simulated_purity, clone_length_ratio)
}
\arguments{
\item{simulated_purity}{Full path to the chromosome fasta directory}

\item{clone_length_ratio}{Ratio of the genome length of the clone with CNA to normal diploid genome}
}
\description{
Deprecated - Function for calculating clone proportion for normal contamination in non-pure simulated tumour
}
\author{
naser.ansari-pour
}
\keyword{internal}
