% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_and_merge_bams.R
\name{contaminate_gain_clone_bam}
\alias{contaminate_gain_clone_bam}
\title{Deprecated - Function for adding normal contamination to a GAIN clone}
\usage{
contaminate_gain_clone_bam(
  simulated_purity,
  downsampling_seed,
  samtools,
  clone,
  normal,
  clone_normal
)
}
\arguments{
\item{simulated_purity}{Full path to the chromosome fasta directory}

\item{downsampling_seed}{The seed value to be used for downsampling with samtools}

\item{samtools}{Full path to the samtools bin}

\item{clone}{Name of the clone with the simulated GAIN}

\item{normal}{Name of the clone with normal diploid copy number}

\item{clone_normal}{Name of the mixed clone to be used for the output BAM}
}
\description{
Deprecated - Function for adding normal contamination to a GAIN clone
}
\author{
naser.ansari-pour
}
\keyword{internal}
