% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2bam_simulate.R
\name{fasta2bam_simulate}
\alias{fasta2bam_simulate}
\title{Function to simulate BAMs from a fasta sequence input}
\usage{
fasta2bam_simulate(
  clone_fasta_filename,
  art_bin,
  haploid_cov,
  read_length,
  fragment_size,
  fragment_size_sd,
  tmp_dir,
  fq1,
  fq2,
  bwa,
  refseq,
  bam,
  ncores,
  samtools,
  logfile,
  skip_art = FALSE,
  skip_bwa = FALSE
)
}
\arguments{
\item{clone_fasta_filename}{The name of the fasta file name for BAM generation}

\item{art_bin}{Full path to the ART bin tool for Illumina read simulation}

\item{haploid_cov}{Sequencing coverage depth to be simulated per chromosome copy (numeric)}

\item{read_length}{Length of the Illumina reads to be simulated (integer, usually 150)}

\item{fragment_size}{Mean size of the sequencing library fragments to be simulated (integer)}

\item{fragment_size_sd}{Standard deviation of the size of the sequencing library fragments to be simulated (integer)}

\item{tmp_dir}{Name of the temporary directory for samtools run (character string, e.g. "TMP")}

\item{fq1}{Name of the fastq (mate 1) filename to be generated by ART (character string)}

\item{fq2}{Name of the fastq (mate 2) filename to be generated by ART (character string)}

\item{bwa}{Full path to the BWA tool for read alignment}

\item{refseq}{Full path to the fasta reference file to be used by BWA}

\item{bam}{Name of the BAM file to be generated (string)}

\item{ncores}{Number of cores to be used in running BWA and samtools}

\item{samtools}{Full path to the samtools bin}

\item{logfile}{Name of the logfile for the BWA run (string)}

\item{skip_art}{set to TRUE if ART run is already complete - files are expected in the working directory (default = FALSE)}

\item{skip_bwa}{set to TRUE if BWA run is already complete - files are expected in the working directory (default = FALSE)}
}
\description{
The function takes a fasta file input (i.e. has 1 to N fasta sequences in the the fasta file depending on copy number status),
simulates Illumina reads using ART and aligns the reads using BWA mem
}
\author{
naser.ansari-pour
}
