% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diploid_bam.R
\name{generate_diploid_bam}
\alias{generate_diploid_bam}
\title{Create Normal diploid BAM for a chromosome}
\usage{
generate_diploid_bam(
  chrom,
  fasta_dir,
  phase_dir,
  art_bin,
  haploid_coverage,
  read_length,
  fragment_size,
  fragment_size_sd,
  tmp_dir,
  bwa,
  refseq,
  bam,
  ncores,
  samtools,
  logfile,
  skip_art = FALSE,
  skip_bwa = FALSE
)
}
\arguments{
\item{chrom}{Name of the chromosome to be simulated}

\item{fasta_dir}{Full path to the chromosome fasta directory}

\item{phase_dir}{Full path to the phase reference file directory}

\item{art_bin}{Full path to the ART bin tool for Illumina read simulation}

\item{haploid_coverage}{Sequencing coverage depth to be simulated per chromosome copy (numeric)}

\item{read_length}{Length of the Illumina reads to be simulated (integer, usually 150)}

\item{fragment_size}{Mean size of the sequencing library fragments to be simulated (integer)}

\item{fragment_size_sd}{Standard deviation of the size of the sequencing library fragments to be simulated (integer)}

\item{tmp_dir}{Name of the temporary directory for samtools run (character string, e.g. "TMP")}

\item{bwa}{Full path to the BWA tool for read alignment}

\item{refseq}{Full path to the fasta reference file to be used by BWA}

\item{bam}{Name of the BAM file to be generated (string)}

\item{ncores}{Number of cores to be used in running BWA and samtools}

\item{samtools}{Full path to the samtools bin}

\item{logfile}{Name of the logfile for the BWA run (string)}

\item{skip_art}{set to TRUE if ART run is already complete - files are expected in the working directory (default = FALSE)}

\item{skip_bwa}{set to TRUE if BWA run is already complete - files are expected in the working directory (default = FALSE)}
}
\description{
The function creates a diploid BAM for a given chromosome mainly for copy number paired-analysis
as germline or to generate a non-CNA chromosome for the tumour
}
\author{
naser.ansari-pour
}
