% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomewide_merge_bam_list.R
\name{genomewide_merge_bam_list}
\alias{genomewide_merge_bam_list}
\title{Function to generate a list of bams to be merged for a genomewide level bam}
\usage{
genomewide_merge_bam_list(chrom_names, simulated_purity, out_dir, samplename)
}
\arguments{
\item{chrom_names}{Name of the chromosomes to be used for merging their respective chromosome-level bams}

\item{simulated_purity}{The tumour purity to be simulated (numeric, ranging (0,1])}

\item{out_dir}{Full path to the out directory where genomewide files will be written}

\item{samplename}{Sample name to be used for the final genomewide bam file with simulated CNA}
}
\description{
Function to generate a list of bams to be merged for a genomewide level bam
}
\author{
naser.ansari-pour
}
