% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_mix.R
\name{loh_mix_prop}
\alias{loh_mix_prop}
\title{Non-piecewise function for calculating the proportion of downsampling from LOH and normal BAMs}
\usage{
loh_mix_prop(cna_ccf, cna_length, chrom_length)
}
\arguments{
\item{cna_ccf}{The tumour purity to be simulated (numeric, ranging (0,1])}

\item{cna_length}{CNA sequence length (integer value)}

\item{chrom_length}{Chromosome sequence length (integer value)}
}
\description{
Correct function to calculate the prop for the LOH bam for subclonal events
}
\author{
naser.ansari-pour
}
\keyword{internal}
