% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_and_merge_bams.R
\name{merge_mini_bams}
\alias{merge_mini_bams}
\title{Function for merging BAM segments/regions to form a full chromosome}
\usage{
merge_mini_bams(samtools, mini_bams_list_file, ncores, clone_bam_filename)
}
\arguments{
\item{samtools}{Full path to the samtools bin}

\item{mini_bams_list_file}{Full path to a one-column file listing the bam names to be merged (including the .bam file extension)}

\item{ncores}{Number of cores to be used in running BWA and samtools}

\item{clone_bam_filename}{The file name (including the .bam file extension) to be used for the final merged bam output}
}
\description{
This functions merges the pieces/segments of the chromosome based on the bam list file
that need to be stitched back together some of which will contain CNA and
some are diploid regions (if any) in between
}
\author{
naser.ansari-pour
}
