% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_and_merge_bams.R
\name{mix_bams_loh}
\alias{mix_bams_loh}
\title{Function for mixing BAMs for subclonal LOH events}
\usage{
mix_bams_loh(
  simulated_purity,
  cna_ccf,
  downsampling_seed,
  samtools,
  clone0,
  clone1,
  clone01
)
}
\arguments{
\item{simulated_purity}{Full path to the chromosome fasta directory}

\item{cna_ccf}{The tumour purity to be simulated (numeric, ranging (0,1])}

\item{downsampling_seed}{The seed value to be used for downsampling with samtools}

\item{samtools}{Full path to the samtools bin}

\item{clone0}{Name of the clone with the simulated LOH}

\item{clone1}{Name of the clone with normal diploid copy number}

\item{clone01}{Name of the mixed clone to be used for the output BAM}
}
\description{
Function for mixing BAMs for subclonal LOH events
}
\author{
naser.ansari-pour
}
