% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlatformeR.R
\name{PlatformeR}
\alias{PlatformeR}
\title{PlatformeR main function for generating per chromosome BAMs (with CNA or diploid)}
\usage{
PlatformeR(
  cna_simulate_file,
  chrom,
  out_dir,
  fasta_dir,
  phase_dir,
  art_bin,
  coverage,
  read_length,
  fragment_size,
  fragment_size_sd,
  tmp_dir,
  bwa,
  ncores,
  samtools,
  loh_haplotype,
  gain_haplotype,
  simulated_purity,
  generate_normal,
  normal_coverage,
  generate_diploid_normal = TRUE
)
}
\arguments{
\item{cna_simulate_file}{Full path to the file with CNA simulations which is a table with five columns (chr, startpos, endpos, cna, CCF) with no headers}

\item{chrom}{Name of the chromosome to be simulated}

\item{out_dir}{Full path to the out directory where genomewide files will be written}

\item{fasta_dir}{Full path to the chromosome fasta directory}

\item{phase_dir}{Full path to the phase reference file directory}

\item{art_bin}{Full path to the ART bin tool for Illumina read simulation}

\item{coverage}{Sequencing coverage depth to be simulated per chromosome copy for the tumour (numeric)}

\item{read_length}{Length of the Illumina reads to be simulated (integer, usually 150)}

\item{fragment_size}{Mean size of the sequencing library fragments to be simulated (integer)}

\item{fragment_size_sd}{Standard deviation of the size of the sequencing library fragments to be simulated (integer)}

\item{tmp_dir}{Name of the temporary directory for samtools run (character string, e.g. "TMP")}

\item{bwa}{Full path to the BWA tool for read alignment}

\item{ncores}{Number of cores to be used in running BWA and samtools}

\item{samtools}{Full path to the samtools bin}

\item{loh_haplotype}{The haplotype onto which LOH should be simulated (string, either 'maternal' or 'paternal')}

\item{gain_haplotype}{The haplotype onto which GAIN should be simulated (string, either 'maternal' or 'paternal')}

\item{simulated_purity}{The tumour purity to be simulated (numeric, ranging (0,1])}

\item{generate_normal}{set to FALSE if a normal germline diploid BAM for the same chromosome is not required e.g. for tumour-only simulations (default=TRUE)}

\item{normal_coverage}{Sequencing coverage depth to be simulated per chromosome copy for the diploid germline/control sample (numeric)}

\item{generate_diploid_normal}{set to FALSE if a normal germline diploid BAM is not required for a non-CNA chromosome in the tumour e.g. for tumour-only simulations (default=TRUE)}
}
\description{
The function creates a BAM for a chromosome with >=1 CNA simulated in it
and creates a diploid germline normal for the same chromosome if required (i.e.paired analysis)
}
\author{
naser.ansari-pour
}
