% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chrom_fasta.R
\name{prepare_chrom_fasta}
\alias{prepare_chrom_fasta}
\title{Function to generate chromosome fasta files}
\usage{
prepare_chrom_fasta(chrom, fasta_dir, phase_dir)
}
\arguments{
\item{chrom}{Name of the chromosome to be simulated (chrom number without 'chr' e.g. 1,2,...,22)}

\item{fasta_dir}{Full path to the chromosome fasta directory}

\item{phase_dir}{Full path to the phase reference file directory}
}
\description{
This function generates maternal and paternal fasta files per chromosome
}
\author{
naser.ansari-pour
}
