% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_regions.R
\name{simulate_gain_region}
\alias{simulate_gain_region}
\title{Function to simulate GAIN region BAM}
\usage{
simulate_gain_region(
  chrom,
  gain_haplotype,
  gain_region,
  gain_simulate,
  art_bin,
  haploid_cov,
  read_length,
  fragment_size,
  fragment_size_sd,
  tmp_dir,
  bwa,
  refseq,
  bam,
  ncores,
  samtools,
  logfile,
  skip_art = FALSE,
  skip_bwa = FALSE
)
}
\arguments{
\item{chrom}{Name of the chromosome to be simulated (chrom number without 'chr' e.g. 1,2,...,22)}

\item{gain_haplotype}{The haplotype onto which GAIN should be simulated (string, either 'maternal' or 'paternal')}

\item{gain_region}{The region number to be simulated among the gain_simulate regions (integer)}

\item{gain_simulate}{A dataframe with two columns containing startpos and endpos of each gain region}

\item{art_bin}{Full path to the ART bin tool for Illumina read simulation}

\item{haploid_cov}{Sequencing coverage depth to be simulated per chromosome copy (numeric)}

\item{read_length}{Length of the Illumina reads to be simulated (integer, usually 150)}

\item{fragment_size}{Mean size of the sequencing library fragments to be simulated (integer)}

\item{fragment_size_sd}{Standard deviation of the size of the sequencing library fragments to be simulated (integer)}

\item{tmp_dir}{Name of the temporary directory for samtools run (character string, e.g. "TMP")}

\item{bwa}{Full path to the BWA tool for read alignment}

\item{refseq}{Full path to the fasta reference file to be used by BWA}

\item{bam}{Name of the BAM file to be generated (string)}

\item{ncores}{Number of cores to be used in running BWA and samtools}

\item{samtools}{Full path to the samtools bin}

\item{logfile}{Name of the logfile for the BWA run (string)}

\item{skip_art}{set to TRUE if ART run is already complete - files are expected in the working directory (default = FALSE)}

\item{skip_bwa}{set to TRUE if BWA run is already complete - files are expected in the working directory (default = FALSE)}
}
\description{
This function simulates a GAIN bam for a given region of interest on a chromosome
}
\author{
naser.ansari-pour
}
